Move array extensions into their own file and test them.
